diary-confluence/requirements.txt
atlassian-python-api==3.20.0
python-dateutil==2.8.2
types-python-dateutil==2.8.15
